requirements.txt
playwright>=1.43.0
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
dnspython>=2.6.0
flask>=3.0.0
